Change interface to new class.  Simple menu for app detection.  Add json file for tracking the support collectd plugin.  Refactor apache code out of conf_colected_plugin into its own class
