+++ config.py
@@ -1,2 +1,3 @@
 # install script will pass in the name of log file
 INSTALL_LOG = None
+DEBUG = True
